+++ src/run.py
@@ -3,6 +3,8 @@ from RatingLib import User, Movie
 from tqdm import tqdm
 from RatingSystem import RatingSystem, RatingSystemCompetition
 from SampleSystems import NaiveRating, AverageMovieRating, GlobalAverageMovieRating, Cheater, AverageUserRating
+from system156007 import MySystem
+from StudentSystems import Sys147715
 # from secret_system import MySystem as secret
 
 def main():
@@ -29,7 +31,8 @@ def main():
     competition.register(AverageMovieRating())
     competition.register(GlobalAverageMovieRating())
     competition.register(Cheater())
-    competition.register(AverageUserRating())    
+    competition.register(AverageUserRating())
+    competition.register(MySystem())
     competition.build_round_robin()
     #run the competition - it prints out the results
     competition.compete()
